+++ src/components/Home/Exercise/ExerciseThumbnail.tsx
@@ -1,5 +1,5 @@
 import { Card } from "@/components/ui/card";
-import { useTrainings } from "@/recoills/tranings";
+import { useTrainings } from "@/recoils/tranings";
 import { Link } from "react-router-dom";
 import styled from "styled-components";
 import Col from "@/components/layouts/Col";
