+++ src/components/Home/More/HomeSectionMore.tsx
@@ -1,10 +1,10 @@
-import { useTrainings } from "@/recoills/tranings";
+import { useTrainings } from "@/recoils/tranings";
 import styled from "styled-components";
 import { Constants } from "@/lib/constant";
 import SheetWrapper from "@/components/layouts/SheetWrapper";
 import Title from "@/components/layouts/Title";
-import { useNews } from "../../../recoills/news";
-import { useNotices } from "../../../recoills/notices";
+import { useNews } from "@/recoils/news";
+import { useNotices } from "@/recoils/notices";
 import MoreCard from "./MoreCard";
 
 const Container = styled.div``;
