+++ src/pages/News/NewsMobile.tsx
@@ -4,7 +4,7 @@ import styled from "styled-components";
 import RightArrow from "@/lib/assets/svgs/arrow_back_ios.svg";
 import MobilePhotoCard from "@/components/Photo/MobilePhotoCard";
 import Loading from "@/components/common/Skeletons/Loading";
-import { useNews } from "../../recoills/news";
+import { useNews } from "@/recoils/news";
 import { NewsParamsType } from "@/lib/types/NewsParamsType";
 
 // TODO: 모바일 헤더 높이 44px 고정 상수화
