+++ src/pages/admin/News/AdminNewsDetail.tsx
@@ -2,7 +2,7 @@ import { useParams } from "react-router-dom";
 import NewsForm from "@/components/admin/form/NewsForm";
 import { Constants } from "@/lib/constant";
 import Title from "@/components/layouts/Title";
-import { useNews } from "../../../recoills/news";
+import { useNews } from "@/recoils/news";
 import { useEffect } from "react";
 
 const AdminNewsDetail = () => {
