+++ src/pages/admin/Notice/AdminNotice.tsx
@@ -3,7 +3,7 @@ import { Link } from "react-router-dom";
 import FormContainer from "@/components/admin/form/FormContainer";
 import { NewArticleButton } from "@/components/admin/form/StyledComponent/FormContainer";
 import ListContainer from "@/components/layouts/ListContainer";
-import { useNotices } from "../../../recoills/notices";
+import { useNotices } from "@/recoils/notices";
 import Row from "@/components/layouts/Row";
 
 const AdminNotice = () => {
