+++ src/pages/admin/Notice/AdminNoticeDetail.tsx
@@ -3,7 +3,7 @@ import { useParams } from "react-router-dom";
 import NoticeForm from "@/components/admin/form/NoticeForm";
 import { Constants } from "@/lib/constant";
 import Title from "@/components/layouts/Title";
-import { useNotices } from "../../../recoills/notices";
+import { useNotices } from "@/recoils/notices";
 
 const AdminNoticeDetail = () => {
   const { id } = useParams();
